Re-enable NetFX in FileSystem
