+++ src/libraries/System.IO.FileSystem/tests/System.IO.FileSystem.Tests.csproj
@@ -1,8 +1,8 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <IncludeRemoteExecutor>true</IncludeRemoteExecutor>
-    <TargetFrameworks>$(NetCoreAppCurrent)-Windows_NT;$(NetCoreAppCurrent)-Unix</TargetFrameworks>
+    <TargetFrameworks>$(NetCoreAppCurrent)-Windows_NT;$(NetCoreAppCurrent)-Unix;$(NetFrameworkCurrent)</TargetFrameworks>
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Base\BaseGetSetAttributes.cs" />
@@ -13,11 +13,12 @@
     <Compile Include="Base\AllGetSetAttributes.cs" />
     <Compile Include="Base\StaticGetSetTimes.cs" />
     <Compile Include="Directory\EnumerableTests.cs" />
+    <Compile Include="Directory\GetLogicalDrives.cs" />
+    <Compile Include="Enumeration\TrimmedPaths.cs" />
     <Compile Include="FileInfo\GetSetAttributesCommon.cs" />
-    <Compile Include="FileInfo\IsReadOnly.cs" />    
+    <Compile Include="FileInfo\IsReadOnly.cs" />
     <Compile Include="FileInfo\Replace.cs" />
     <Compile Include="FileStream\Handle.cs" />
-    <Compile Include="Directory\GetLogicalDrives.cs" />
     <Compile Include="FileStream\EndRead.cs" />
     <Compile Include="FileStream\EndWrite.cs" />
     <Compile Include="FileStream\LockUnlock.cs" />
@@ -25,31 +26,43 @@
     <Compile Include="File\EncryptDecrypt.cs" />
     <Compile Include="File\GetSetAttributesCommon.cs" />
     <Compile Include="File\Replace.cs" />
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetsNetCoreApp)' == 'true'">
+    <Compile Include="Directory\EnumerableTests.netcoreapp.cs" />
+    <Compile Include="Enumeration\AttributeTests.cs" />
+    <Compile Include="Enumeration\ConstructionTests.cs" />
+    <Compile Include="Enumeration\ErrorHandlingTests.cs" />
+    <Compile Include="Enumeration\ExampleTests.cs" />
+    <Compile Include="Enumeration\FileSystemNameTests.cs" />
+    <Compile Include="Enumeration\IncludePredicateTests.cs" />
+    <Compile Include="Enumeration\MatchCasingTests.cs" />
+    <Compile Include="Enumeration\MatchTypesTests.cs" />
+    <Compile Include="Enumeration\PatternTransformTests.cs" />
+    <Compile Include="Enumeration\RemovedDirectoryTests.cs" />
+    <Compile Include="Enumeration\RootTests.cs" />
+    <Compile Include="Enumeration\SkipAttributeTests.cs" />
+    <Compile Include="Enumeration\SpecialDirectoryTests.cs" />
+    <Compile Include="Enumeration\TrimmedPaths.netcoreapp.cs" />
     <Compile Include="File\AppendAsync.cs" />
+    <Compile Include="File\Move.netcoreapp.cs" />
     <Compile Include="File\ReadWriteAllLinesAsync.cs" />
     <Compile Include="File\ReadWriteAllBytesAsync.cs" />
     <Compile Include="File\ReadWriteAllTextAsync.cs" />
+    <Compile Include="FileInfo\MoveTo.netcoreapp.cs" />
+    <Compile Include="FileStream\ctor_sfh_fa_buffer_async.cs" />
+    <Compile Include="FileStream\ctor_str_fm_fa_fs_buffer_async.cs" />
+    <Compile Include="FileStream\CopyToAsync.cs" />
     <Compile Include="FileStream\DisposeAsync.cs" />
+    <Compile Include="FileStream\FlushAsync.cs" />
+    <Compile Include="FileStream\IsAsync.cs" />
+    <Compile Include="FileStream\ReadAsync.cs" />
     <Compile Include="FileStream\ReadWriteSpan.cs" />
-    <Compile Include="Enumeration\ConstructionTests.cs" />
-    <Compile Include="Enumeration\SpecialDirectoryTests.cs" />
-    <Compile Include="Enumeration\SkipAttributeTests.cs" />
-    <Compile Include="Enumeration\FileSystemNameTests.cs" />
-    <Compile Include="Enumeration\MatchCasingTests.cs" />
-    <Compile Include="Enumeration\TrimmedPaths.cs" />
-    <Compile Include="Enumeration\ErrorHandlingTests.cs" />
-    <Compile Include="Enumeration\IncludePredicateTests.cs" />
-    <Compile Include="Enumeration\PatternTransformTests.cs" />
-    <Compile Include="Enumeration\RootTests.cs" />
-    <Compile Include="Enumeration\AttributeTests.cs" />
-    <Compile Include="Enumeration\MatchTypesTests.cs" />
-    <Compile Include="Enumeration\ExampleTests.cs" />
-    <Compile Include="Enumeration\RemovedDirectoryTests.cs" />
+    <Compile Include="FileStream\WriteAsync.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetsUnix)' == 'true'">
     <Compile Include="FileSystemTest.Unix.cs" />
   </ItemGroup>
-  <ItemGroup Condition="'$(TargetsWindows)' == 'true'">
+  <ItemGroup Condition="'$(TargetsWindows)' == 'true' or '$(TargetFramework)' == '$(NetFrameworkCurrent)'">
     <Compile Include="FileSystemTest.Windows.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -66,38 +79,31 @@
     <Compile Include="FileInfo\GetSetAttributes.cs" />
     <Compile Include="FileInfo\Length.cs" />
     <Compile Include="FileInfo\Open.cs" />
-    <Compile Include="FileStream\FlushAsync.cs" />
     <Compile Include="FileStream\Pipes.cs" />
     <Compile Include="FileStream\SafeFileHandle.cs" />
-    <Compile Include="FileStream\IsAsync.cs" />
     <Compile Include="FileStream\CanTimeout.cs" />
     <Compile Include="FileStream\CanSeek.cs" />
     <Compile Include="FileStream\Name.cs" />
     <Compile Include="FileStream\CanWrite.cs" />
     <Compile Include="FileStream\CanRead.cs" />
-    <Compile Include="FileStream\CopyToAsync.cs" />
     <Compile Include="FileStream\Buffering_regression.cs" />
     <Compile Include="FileStream\Flush.cs" />
     <Compile Include="FileStream\Dispose.cs" />
-    <Compile Include="FileStream\WriteAsync.cs" />
     <Compile Include="FileStream\Write.cs" />
     <Compile Include="FileStream\ToString.cs" />
     <Compile Include="FileStream\WriteByte.cs" />
-    <Compile Include="FileStream\ReadAsync.cs" />
     <Compile Include="FileStream\Read.cs" />
     <Compile Include="FileStream\ReadByte.cs" />
     <Compile Include="FileStream\SetLength.cs" />
     <Compile Include="FileStream\Length.cs" />
     <Compile Include="FileStream\Position.cs" />
     <Compile Include="FileStream\Seek.cs" />
-    <Compile Include="FileStream\ctor_sfh_fa_buffer_async.cs" />
     <Compile Include="FileStream\ctor_sfh_fa_buffer.cs" />
     <Compile Include="FileStream\ctor_sfh_fa.cs" />
     <Compile Include="FileStream\ctor_str_fm_fa_fs.write.cs" />
     <Compile Include="FileStream\ctor_str_fm_fa_fs.read.cs" />
     <Compile Include="FileStream\ctor_str_fm_fa_fs.delete.cs" />
     <Compile Include="FileStream\ctor_str_fm_fa_fs_buffer_fo.cs" />
-    <Compile Include="FileStream\ctor_str_fm_fa_fs_buffer_async.cs" />
     <Compile Include="FileStream\ctor_str_fm_fa_fs_buffer.cs" />
     <Compile Include="FileStream\ctor_str_fm_fa_fs.cs" />
     <Compile Include="FileStream\ctor_str_fm_fa.cs" />
@@ -113,12 +119,6 @@
     <Compile Include="UnseekableFileStream.cs" />
     <Compile Include="FSAssert.cs" />
     <!-- Ported -->
-    <Compile Include="PortedCommon\CommonUtilities.cs" />
-    <Compile Include="PortedCommon\DllImports.cs" />
-    <Compile Include="PortedCommon\IOInputs.cs" />
-    <Compile Include="PortedCommon\IOServices.cs" />
-    <Compile Include="PortedCommon\PathInfo.cs" />
-    <Compile Include="PortedCommon\ReparsePointUtilities.cs" />
     <Compile Include="DirectoryInfo\EnumerableAPIs.cs" />
     <Compile Include="DirectoryInfo\GetDirectories.cs" />
     <Compile Include="DirectoryInfo\GetFiles.cs" />
@@ -160,6 +160,12 @@
     <Compile Include="FileInfo\ToString.cs" />
     <Compile Include="FileInfo\AppendText.cs" />
     <Compile Include="FileInfo\CopyTo.cs" />
+    <Compile Include="PortedCommon\CommonUtilities.cs" />
+    <Compile Include="PortedCommon\DllImports.cs" />
+    <Compile Include="PortedCommon\IOInputs.cs" />
+    <Compile Include="PortedCommon\IOServices.cs" />
+    <Compile Include="PortedCommon\PathInfo.cs" />
+    <Compile Include="PortedCommon\ReparsePointUtilities.cs" />
     <!-- Helpers -->
     <Compile Include="$(CommonPath)Interop\Windows\Kernel32\Interop.GenericOperations.cs" Link="Interop\Windows\Interop.GenericOperations.cs" />
     <Compile Include="$(CommonTestPath)System\Buffers\NativeMemoryManager.cs">
